clean up ghost code
