+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/HardwarePushbot.java
@@ -2,10 +2,6 @@ package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.TouchSensor;
-import com.qualcomm.hardware.rev.Rev2mDistanceSensor;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.hardware.rev.RevColorSensorV3;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -58,13 +54,6 @@ public class HardwarePushbot {
         public DcMotor  rightFrontDriveWheel  = null; //motor for right front wheel
         public DcMotor  rightBackDriveWheel  = null; //motor for right back wheel
 
-        //public Rev2mDistanceSensor distanceSensor = null;
-        //public TouchSensor touchSensor = null;
-
-        public DcMotor armMotor = null;
-        //public DcMotor verticalMotor = null;
-        public Servo clawServo = null;
-        public DcMotor rotateMotor = null;
         
         /* local OpMode members. */
         HardwareMap hwMap           =  null;
@@ -73,7 +62,6 @@ public class HardwarePushbot {
         /* Constructor */
         public HardwarePushbot()
         {
-
         }
 
         /* Initialize standard Hardware interfaces */
@@ -87,33 +75,18 @@ public class HardwarePushbot {
             leftBackDriveWheel  = hwMap.get(DcMotor.class, "leftBackDriveWheel");
             rightFrontDriveWheel = hwMap.get(DcMotor.class, "rightFrontDriveWheel");
             rightBackDriveWheel = hwMap.get(DcMotor.class, "rightBackDriveWheel");
-            
-            armMotor = hwMap.get(DcMotor.class, "armMotor");
-            //verticalMotor = hwMap.get(DcMotor.class, "verticalMotor");
-            rotateMotor = hwMap.get(DcMotor.class, "rotateMotor");
-            
-            clawServo = hwMap.get(Servo.class,"clawServo");
-            //Define and initialize distance sensor
-            //distanceSensor = hwMap.get(Rev2mDistanceSensor.class, "distanceSensor");
-            //touchSensor = hwMap.get(TouchSensor.class, "touchSensor");
 
+            //left wheels are counterclockwise and because of 90 degree gearboxes, they all are forward
             leftFrontDriveWheel.setDirection(DcMotor.Direction.FORWARD);// Set to REVERSE if using AndyMark motors
             leftBackDriveWheel.setDirection(DcMotor.Direction.FORWARD);
-            rightFrontDriveWheel.setDirection(DcMotor.Direction.REVERSE);
-            rightBackDriveWheel.setDirection(DcMotor.Direction.REVERSE);// Set to FORWARD if using AndyMark motors
-            
-            armMotor.setDirection(DcMotor.Direction.FORWARD);
-            rotateMotor.setDirection(DcMotor.Direction.FORWARD);
+            rightFrontDriveWheel.setDirection(DcMotor.Direction.FORWARD);
+            rightBackDriveWheel.setDirection(DcMotor.Direction.FORWARD);// Set to FORWARD if using AndyMark motors
 
             // Set all motors to zero power
             leftFrontDriveWheel.setPower(0);
             leftBackDriveWheel.setPower(0);
             rightFrontDriveWheel.setPower(0);
             rightBackDriveWheel.setPower(0);
-            
-            armMotor.setPower(0);
-            rotateMotor.setPower(0);
-
 
             // Set all motors to run without encoders.
             // May want to use RUN_USING_ENCODERS if encoders are installed.
@@ -121,23 +94,6 @@ public class HardwarePushbot {
             leftBackDriveWheel.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
             rightFrontDriveWheel.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
             rightBackDriveWheel.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-            
-            armMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            // verticalMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            rotateMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODERS);
-            
-            rotateMotor.setZeroPowerBehavior(DcMotorEx.ZeroPowerBehavior.BRAKE);
-            // armMotor.setZeroPowerBehavior(DcMotorEx.ZeroPowerBehavior.BRAKE);
-            armMotor.setZeroPowerBehavior(DcMotorEx.ZeroPowerBehavior.BRAKE);
-
-            // Define and initialize ALL installed servos.
-            // leftClaw  = hwMap.get(Servo.class, "left_hand");
-            //grip = hwMap.get(Servo.class, "grip");
-            // leftClaw.setPosition(MID_SERVO);
-            //grip.setPosition(MID_SERVO);
-            
-            //camera = hwMap.get(Servo.class, "camera");
-            //camera.setPosition(MID_SERVO);
         }
     }
